requirements.txt
faiss-cpu
sentence-transformers
transformers
playwright
numpy
pytest
typing-extensions
